requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
python-frontmatter>=1.1,<2.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
neo4j>=5.18,<6.0
langchain>=0.2,<0.4
langchain-openai>=0.1,<0.3
langgraph>=0.1,<0.4
openai>=1.30,<2.0
pypdf>=4.2,<6.0
pyyaml>=6.0,<7.0
